+++ connectorx/tests/test_mssql.rs
@@ -112,7 +112,7 @@ pub fn verify_arrow_results(result: Vec<RecordBatch>) {
                     .as_any()
                     .downcast_ref::<Float64Array>()
                     .unwrap()
-                    .eq(&Float64Array::from(vec![None, Some(3.1 as f64)])));
+                    .eq(&Float64Array::from(vec![None, Some(3.1_f64)])));
 
                 assert!(rb
                     .column(4)
@@ -154,10 +154,10 @@ pub fn verify_arrow_results(result: Vec<RecordBatch>) {
                     .downcast_ref::<Float64Array>()
                     .unwrap()
                     .eq(&Float64Array::from(vec![
-                        Some(2.2 as f64),
-                        Some(3 as f64),
-                        Some(7.8 as f64),
-                        Some(-10 as f64),
+                        Some(2.2_f64),
+                        Some(3_f64),
+                        Some(7.8_f64),
+                        Some(-10_f64),
                     ])));
 
                 assert!(rb
